requirements.txt
# 核心依赖
Pillow>=10.0.0
piexif>=1.1.3
gpxpy>=1.5.0
pandas>=2.0.0
python-dateutil>=2.8.2

# UI框架
nicegui>=1.4.0

